+++ projects/01_task_board_cli/PFCompactRenderer.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+#import "PFTaskRendering.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface PFCompactRenderer : NSObject <PFTaskRendering>
+@end
+
+NS_ASSUME_NONNULL_END