+++ projects/01_task_board_cli/PFConsoleRenderer.h
@@ -1,7 +1,11 @@
 #import <Foundation/Foundation.h>
-#import "PFTaskBoard.h"
+#import "PFTaskRendering.h"
 
-@interface PFConsoleRenderer : NSObject
+NS_ASSUME_NONNULL_BEGIN
+
+@interface PFConsoleRenderer : NSObject <PFTaskRendering>
 - (void)renderBoard:(PFTaskBoard * _Nonnull)board;
 - (void)renderBoard:(PFTaskBoard * _Nonnull)board onlyState:(nullable NSNumber *)stateFilter;
 @end
+
+NS_ASSUME_NONNULL_END