+++ projects/01_task_board_cli/PFTaskRendering.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import "PFTaskBoard.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol PFTaskRendering <NSObject>
+- (void)renderBoard:(PFTaskBoard *)board onlyState:(nullable NSNumber *)stateFilter;
+@end
+
+NS_ASSUME_NONNULL_END